+++ src/model/jms.cpp
@@ -226,10 +226,13 @@ namespace Invader {
         Marker m;
         m.name = string_from_string(cursor, true);
         m.region = read_next_uint32(cursor);
-        if((m.region & NULL_INDEX) == NULL_INDEX) {
-            m.region = 0; // -1 = 0 I guess
+        if(m.region == NULL_INDEX) {
+            m.region = 0;
         }
         m.node = read_next_uint32(cursor);
+        if(m.node == NULL_INDEX) {
+            m.node = 0;
+        }
         m.rotation = quaternion_from_string(cursor);
         m.position = point3d_from_string(cursor) / 100.0F;
         m.radius = read_next_float(cursor);
@@ -318,10 +321,13 @@ namespace Invader {
         SET_CURSOR
         Triangle t;
         t.region = read_next_uint32(cursor);
-        if((t.region & NULL_INDEX) == NULL_INDEX) {
+        if(t.region == NULL_INDEX) {
             t.region = 0; // -1 = 0 I guess
         }
         t.shader = read_next_uint32(cursor);
+        if(t.shader == NULL_INDEX) {
+            t.shader = 0;
+        }
         t.vertices[0] = read_next_uint32(cursor);
         t.vertices[2] = read_next_uint32(cursor);
         t.vertices[1] = read_next_uint32(cursor);